Create useArchetypeMutations hook to consolidate archetype mutations
